Some changes and moves. Now working and some refactorings. Added some output in App.

+++ src/main/java/org/example/model/Product.java
@@ -6,8 +6,8 @@ public interface Product {
      int getId();
      String getProductName();
      void setProductName(String productName);
-     double getPrice();
-     void setPrice(double price);
+     int getPrice();
+     void setPrice(int price);
